Teams can now be eliminated in the middle of a game.
Change the Player Death system to allow for perma deaths in the middle of the game

Respawns now cost 1 Game Score

Added a attempt to respawnTeamMembers when gamescore increases

Wins are now determined when a team is deprecated.

Teams that are deprecated have no chance of reverting back. (Temporary change)

Added Vaporization Checks also known as Game Score thresholds to force teams who are far behind to catch up or be eliminated.
This Vaporization threshold is shown to players who fall below it

Added new PlayerState

Queue now only needs 75% to ready before the game starts

Change Craft for Team Portal to be cheaper